split merge util
